+++ test/system/external_test.rb
@@ -14,6 +14,11 @@ class ExternalTest < ApplicationSystemTestCase
     screenshot("visit-contact-page")
   end
 
+  test "visiting the privacy policy page" do
+    visit privacy_policy_url
+    screenshot("visit-privacy-policy-page")
+  end
+
   test "visit version page" do
     visit version_url
     screenshot("visit-version-page")